rename: main, sign 최상위 뷰 파일 이동 및 index로 이름 변경
