+++ src/js/views/sign/index.ts
@@ -1,10 +1,10 @@
-import { URL } from '../constants/appContants';
-import { signTemplate } from '../templates/sign/signTemplate';
-import { $ } from '../utils/common';
-import { SELECTOR } from '../constants/viewConstants';
-import SignInView from './sign/signInView';
-import SignUpView from './sign/signUpView';
-import EditProfileView from './sign/editProfileView';
+import { URL } from '../../constants/appContants';
+import { signTemplate } from '../../templates/sign/signTemplate';
+import { $ } from '../../utils/common';
+import { SELECTOR } from '../../constants/viewConstants';
+import SignInView from './signInView';
+import SignUpView from './signUpView';
+import EditProfileView from './editProfileView';
 
 export default class SignView {
   private $app: HTMLElement;
